picasso/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.2
django-dramatiq==0.11.6
djangorestframework==3.14.0
dramatiq==1.16.0
drf-yasg==1.21.7
flake8==7.0.0
inflection==0.5.1
kombu==5.3.5
mccabe==0.7.0
packaging==23.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pycodestyle==2.11.1
pyflakes==3.2.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
